bug: fix issue with check moves
Fixed a major issue with the valid moves
Previously, moves that put your in king in check were calculated as valid
Fixed by filtering moves selectively by seeing if they are valid when executed
Move execution checks for putting the king in check and does not allow it.

Currently, Kings in proximity of other kings is still being allowed, working on a fix for that.