bug: fix checks calculation
Previously, checks would stop looking when the first check was found.
There can possibly be many checks when looking for moves that cause check.
Therefore, made the checkForChecks function return an array of all the possible
checks and do it that way.

Possibly refactor later to white and black check only